requirements.txt
Django
python-dotenv
termcolor
pytest
pytest-cov
psycopg2
psycopg2-binary
pylint-django
Pillow

asgiref
sqlparse
django-unfold
